adjusted soem things to look better
