+++ app/src/main/res/layout/progress_spinner_overlay.xml
@@ -6,8 +6,7 @@
     android:layout_height="match_parent"
     android:alpha="0.75"
     android:background="@color/black"
-    android:animateLayoutChanges="true"
-    android:visibility="gone">
+    android:animateLayoutChanges="true">
 
     <ProgressBar
         android:id="@+id/progressSpinner"
